\name{baseballData}
\docType{data}
\alias{baseballData}
\title{Baseball player performance and pay dataset}
\description{
  This data set is a collection of performance statistics 
  for a collection of baseball players including a \code{Salary} variable.
}
\usage{data(baseballData)}
\format{A data frame consisting of 11 variables (columns) for 322 players (rows).}
\source{CRAN: \code{corrgram} package }
\references{
This dataset is a subset of the \code{baseball} dataset from the \code{corrgram} package.

Author: Kevin Wright, 
Copyright: 2006 Kevin Wright, 
License: GPL2.

}
\keyword{datasets}